logwrap: add ChanLogger functionality
There are cases where a logger is dedicated to an object's main
goroutine, but that goroutine spawns additional goroutines to perform
work.  Logging within the subordinate goroutines should be coupled to
the main goroutine, but loggers are generally not safe for concurrent
use.

A ChanLogger packages up the priority and other log information and
sends it to a channel.  The object main goroutine can add that channel
to the ones it monitors, emitting the messages in its context when
they are received.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>